M2: coordinated overhang grading fix (classic path, byte-faithful)
Replace the midpoint-band approximation with the faithful BambuStudio classic
overhang flow and route 100%-overhang to bridge/overhang walls:

(a) Geometry: perimeter_generator.rs traverse_loops now follows
    PerimeterGenerator.cpp:374-432 exactly — bbox+SCALED_EPSILON, clip lower
    series front/back to bbox, inside=intersection_pl_2, remain=diff_pl_2,
    middle=diff_pl_2(inside,front), zero=intersection_pl_2(inside,front),
    classic detect_overhang_degree on middle, chain_and_reorder at end.
    Uses clipper2_utils::{intersection_pl_2,diff_pl_2} (raw clipper2c FFI that
    recovers solution_open) — the clipper_utils safe-wrapper versions DROP
    solution_open and silently return empty open-path clips.
(b) detect_bridge_wall ported (PerimeterGenerator.cpp:241-267): remain_polines ->
    erOverhangPerimeter at overhang_flow, degree 5 (curved) / 6 (straight bridge).
    overhang_flow wired in layer.rs via bridging_flow(frPerimeter, thick_bridges).
(c) ExtrusionPath.overhang_degree i32 -> f64 (ExtrusionEntity.hpp:216 double) so
    GCode::get_overhang_degree_corr_speed interpolates between speed buckets
    (GCode.cpp:5942-5962). ~16 sites; exporter overhang_degree_corr_speed takes
    f64 with the 0->normal fallback only in the interpolation branch + caller guard.
(d) Exporter: per-path ';FEATURE: <role>' emitted on role change (GCode.cpp:6599)
    in BOTH the smooth-speed and plain loop paths; erOverhangPerimeter speed
    (deg5->overhang_totally, deg6->bridge) in get_path_speed (GCode.cpp:5401-5408).

Result (full Benchy compare): Overhang wall 0 -> 90 (= native 90, EXACT);
Outer wall 679->753 (native 772); Inner wall 652->669 (native 667); degrees now
continuous (multiples of 0.1) matching native CLASSIC_DOD; 104 bridge/overhang
paths (deg5:77 deg6:27) vs native 107 (78/29). Filament stable 3795.34.

+++ crates/libslic3r-rs/src/variable_width.rs
@@ -234,8 +234,8 @@ pub fn thick_polyline_to_multi_path(
     }
     // VariableWidth.cpp:93
     if path.polyline.is_valid() {
-        // VariableWidth.cpp:94-95
-        path.overhang_degree = overhang as i32;
+        // VariableWidth.cpp:94 — path.overhang_degree = overhang; (double)
+        path.overhang_degree = overhang;
         multi_path.paths.push(path);
     }
     // VariableWidth.cpp:97
